# [Glow] Add numerical equivalence check to GraphOptzTest.optimizeBatchNormAfterConvButConvReused test case