Summary:
Add numerical equivalence check to GraphOptzTest.optimizeBatchNormAfterConvButConvReused test case

Documentation:

[Optional Fixes #issue]

Test Plan:
ninja test
`ankitgureja-mbp:build_Debug ankitgureja$ ninja test
[0/1] Running tests...
Test project /Users/ankitgureja/build_Debug
      Start  1: BasicIRTest
 1/47 Test  #1: BasicIRTest ....................................   Passed    0.08 sec
      Start  2: Caffe2ImporterTest
 2/47 Test  #2: Caffe2ImporterTest .............................   Passed    2.32 sec
      Start  3: ThreadPoolExecutorTest
 3/47 Test  #3: ThreadPoolExecutorTest .........................   Passed    1.69 sec
      Start  4: Float16Test
 4/47 Test  #4: Float16Test ....................................   Passed    0.02 sec
      Start  5: GemmTest
 5/47 Test  #5: GemmTest .......................................   Passed    0.40 sec
      Start  6: GlowOnnxifiManagerTest
 6/47 Test  #6: GlowOnnxifiManagerTest .........................   Passed    0.38 sec
      Start  7: GraphGradTest
 7/47 Test  #7: GraphGradTest ..................................   Passed    0.40 sec
      Start  8: GraphOptzTest
 8/47 Test  #8: GraphOptzTest ..................................   Passed    0.47 sec
      Start  9: GraphSchedulerTest
 9/47 Test  #9: GraphSchedulerTest .............................   Passed    0.04 sec
      Start 10: GraphTest
10/47 Test #10: GraphTest ......................................   Passed    0.68 sec
      Start 11: HostManagerTest
11/47 Test #11: HostManagerTest ................................   Passed    2.00 sec
      Start 12: HyphenTest
12/47 Test #12: HyphenTest .....................................   Passed    1.94 sec
      Start 13: IROptTest
13/47 Test #13: IROptTest ......................................   Passed    0.06 sec
      Start 14: ImageTest
14/47 Test #14: ImageTest ......................................   Passed    0.19 sec
      Start 15: LLVMIRGenTest
15/47 Test #15: LLVMIRGenTest ..................................   Passed    0.24 sec
      Start 16: MemoryAllocatorTest
16/47 Test #16: MemoryAllocatorTest ............................   Passed    0.05 sec
      Start 17: OnnxImporterTest
17/47 Test #17: OnnxImporterTest ...............................   Passed    0.80 sec
      Start 18: OnnxExporterTest
18/47 Test #18: OnnxExporterTest ...............................   Passed    0.42 sec
      Start 19: CPUBackendCorrectnessTest
19/47 Test #19: CPUBackendCorrectnessTest ......................   Passed   14.37 sec
      Start 20: CPUBackendTest
20/47 Test #20: CPUBackendTest .................................   Passed    0.57 sec
      Start 21: CPUDeviceManagerTest
21/47 Test #21: CPUDeviceManagerTest ...........................   Passed    0.84 sec
      Start 22: CPUGradCheckTest
22/47 Test #22: CPUGradCheckTest ...............................   Passed   10.52 sec
      Start 23: CPUMLTest
23/47 Test #23: CPUMLTest ......................................   Passed   13.52 sec
      Start 24: CPUOperatorGradTest
24/47 Test #24: CPUOperatorGradTest ............................   Passed    0.68 sec
      Start 25: CPUOperatorTest
25/47 Test #25: CPUOperatorTest ................................   Passed   15.77 sec
      Start 26: CPURecommendationSystemTest
26/47 Test #26: CPURecommendationSystemTest ....................   Passed   71.04 sec
      Start 27: CPUTraceEventsTest
27/47 Test #27: CPUTraceEventsTest .............................   Passed    4.10 sec
      Start 28: CPUTypeAToTypeBFunctionConverterTest
28/47 Test #28: CPUTypeAToTypeBFunctionConverterTest ...........   Passed    0.32 sec
      Start 29: InterpreterBackendCorrectnessTest
29/47 Test #29: InterpreterBackendCorrectnessTest ..............   Passed    8.02 sec
      Start 30: InterpreterBackendTest
30/47 Test #30: InterpreterBackendTest .........................   Passed    0.57 sec
      Start 31: InterpreterDeviceManagerTest
31/47 Test #31: InterpreterDeviceManagerTest ...................   Passed    0.46 sec
      Start 32: InterpreterGradCheckTest
32/47 Test #32: InterpreterGradCheckTest .......................   Passed    3.10 sec
      Start 33: InterpreterMLTest
33/47 Test #33: InterpreterMLTest ..............................   Passed   11.54 sec
      Start 34: InterpreterOperatorGradTest
34/47 Test #34: InterpreterOperatorGradTest ....................   Passed    0.32 sec
      Start 35: InterpreterOperatorTest
35/47 Test #35: InterpreterOperatorTest ........................   Passed    3.46 sec
      Start 36: InterpreterRecommendationSystemTest
36/47 Test #36: InterpreterRecommendationSystemTest ............   Passed  196.69 sec
      Start 37: InterpreterTraceEventsTest
37/47 Test #37: InterpreterTraceEventsTest .....................   Passed    2.84 sec
      Start 38: InterpreterTypeAToTypeBFunctionConverterTest
38/47 Test #38: InterpreterTypeAToTypeBFunctionConverterTest ...   Passed    0.33 sec
      Start 39: PartitionerTest
39/47 Test #39: PartitionerTest ................................   Passed    0.38 sec
      Start 40: ProvisionerTest
40/47 Test #40: ProvisionerTest ................................   Passed    0.96 sec
      Start 41: QuantizationTest
41/47 Test #41: QuantizationTest ...............................   Passed    7.77 sec
      Start 42: StatsExporterTest
42/47 Test #42: StatsExporterTest ..............................   Passed    0.37 sec
      Start 43: TensorsTest
43/47 Test #43: TensorsTest ....................................   Passed    0.12 sec
      Start 44: TensorPoolTest
44/47 Test #44: TensorPoolTest .................................   Passed    0.04 sec
      Start 45: ThreadPoolTest
45/47 Test #45: ThreadPoolTest .................................   Passed    0.02 sec
      Start 46: UtilsTest
46/47 Test #46: UtilsTest ......................................   Passed    0.03 sec
      Start 47: ErrorTest
47/47 Test #47: ErrorTest ......................................   Passed    0.02 sec

100% tests passed, 0 tests failed out of 47

Label Time Summary:
EXPENSIVE    = 267.73 sec*proc (2 tests)

Total Test time (real) = 380.98 sec
`